backend/orders/serializers/view.py
from rest_framework import serializers

from orders.models import Order


class OrderSerializer(serializers.ModelSerializer):

    class Meta:
        model = Order
        fields = ('id', 'order_number', 'price_usd', 'delivery_date', 'price_rub')
